design,format/mainpage_treeInfos/디자인 수정 및 파일구조 변경
