+++ src/pages/MainPage/style.js
@@ -1,11 +1,10 @@
 import styled from 'styled-components'
 
 export const Container = styled.main`
+  position: relative;
   width: 421px;
   height: 852px;
-  padding: 24px 14px 14px 0;
   border: 1px solid black;
-  position: relative;
 `
 
 export const InfoSection = styled.section`
@@ -14,7 +13,7 @@ export const InfoSection = styled.section`
   height: auto;
   z-index: 2;
   overflow: hidden;
-  bottom: 0;
-  left: 50%;
+  bottom: 290px;
+  left: 51%;
   transform: translate(-50%, -80%);
 `
